Reset Password initial template updating html template (temporary changes).
